# openapi.yaml
openapi: 3.0.3
info:
  title: Personal Assistant API
  version: 1.0.0
  description: |
    A RESTful API for a personal assistant AI agent that helps with research, content creation,
    and other tasks. The API provides user management, chat functionality, and background job processing.

    ## Authentication
    Most endpoints require authentication using JWT Bearer tokens obtained from the login endpoint.

    ## Rate Limiting
    API requests may be rate-limited to ensure fair usage.

  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /memory/export:
    get:
      summary: Export all memories
      description: |
        Exports all memories for the authenticated user as a ZIP archive.
        The archive contains one JSONL file per memory category:
        - `chat.json`: chat-scoped memories
        - `user.json`: user-scoped unpinned memories
        - `personality-{uuid}.json`: one file per personality with pinned memories

        Each file contains one JSON object per line (JSONL format). Records are
        ordered by `created_at` ascending within each file.
      tags: [Memory Management]
      responses:
        '200':
          description: Memories exported successfully
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error (response may be partial if headers were already sent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /account/export:
    post:
      summary: Enqueue a full-account export
      description: |
        Enqueues an asynchronous full-account export for the authenticated user, built in-process. The
        export is a **ZIP** containing conversations (in the round-trippable Claude `conversations.json`
        shape), personalities (with scratchpads), memories (as a nested `memories.zip`), a
        `files/manifest.json` inventory of the user's uploaded files, and a `manifest.json`.

        Returns `202` with a Job; poll `GET /account/export/{id}` for the phase. **The download link is
        delivered only by email — never in the API response.** This is a deliberate security control:
        access to the app account alone cannot exfiltrate the full account, since the download also
        requires access to the user's mailbox.

        Rate limited to one export per user per 10 minutes.
      tags: [Account Export]
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Export enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: An export was requested too recently
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /account/export/{id}:
    get:
      summary: Get account export status
      description: |
        Returns the export Job. Its `progress` field is a JSON-encoded `AccountExportProgress`
        (`phase` of queued/building/uploading/complete/failed, section `counts`, and a user-safe
        `message`). It never contains the download link — that is emailed.
      tags: [Account Export]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Export job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Export not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /chat:
    get:
      summary: List chats
      description: |
        Returns a paginated list of chats for the current user, ordered by most recent activity.
        By default only **active** (non-archived) threads are returned. Pass `archived=true` to list archived threads only.
      tags: [Chat Management]
      parameters:
        - name: page
          in: query
          description: Page number (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page (default 10)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: name
          in: query
          description: Filter by chat name
          schema:
            type: string
        - name: search
          in: query
          description: Search in chat names and tags
          schema:
            type: string
        - name: tag
          in: query
          description: Filter by exact chat tag
          schema:
            type: string
        - name: personality_id
          in: query
          description: Filter by associated personality ID
          schema:
            type: string
            format: uuid
        - name: is_favorite
          in: query
          description: Filter by favorite status
          schema:
            type: boolean
        - name: archived
          in: query
          description: When true, return only archived threads. When false or omitted, return only active (non-archived) threads.
          schema:
            type: boolean
        - name: min_date
          in: query
          description: Filter chats created after this date
          schema:
            type: string
            format: date-time
        - name: max_date
          in: query
          description: Filter chats created before this date
          schema:
            type: string
            format: date-time
        - name: source
          in: query
          description: Filter to chats imported from this source (e.g. 'openai', 'anthropic'). Used by the post-import thread picker.
          schema:
            type: string
            enum: [openai, anthropic]
      responses:
        '200':
          description: Chats retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/Chat'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create chat
      description: Creates a new chat session. Optional `personality_id` and `model_id` let callers start a thread with a specific persona/model in one round trip; both must be owned by the authenticated user (model must be active). When omitted, the user's default preferences are used.
      tags: [Chat Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the chat session
                  maxLength: 200
                last_message_time:
                  type: string
                  format: date-time
                  description: Timestamp of the most recent message (optional)
                personality_id:
                  type: string
                  format: uuid
                  description: Optional. Personality to associate with the new chat. Must be owned by the user.
                model_id:
                  type: string
                  format: uuid
                  description: Optional. Active model to associate with the new chat.
                tags:
                  type: array
                  maxItems: 10
                  items:
                    type: string
                    maxLength: 10
                  description: Optional chat tags (up to 10 tags, 10 chars each)
                is_favorite:
                  type: boolean
                  description: Whether the chat is marked as favorite
                tts_autoplay_override:
                  type: boolean
                  nullable: true
                  description: Optional per-chat TTS autoplay override. Null/omitted means inherit personality default.
              required:
                - name
      responses:
        '201':
          description: Chat created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Personality or model not found / not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /chat/{id}:
    get:
      summary: Get chat by ID
      description: Returns a specific chat by its ID
      tags: [Chat Management]
      parameters:
        - name: id
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chat retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update chat
      description: Updates an existing chat
      tags: [Chat Management]
      parameters:
        - name: id
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the chat session
                  maxLength: 200
                last_message_time:
                  type: string
                  format: date-time
                  description: Timestamp of the most recent message
                model_id:
                  type: string
                  format: uuid
                  description: Optional model ID to set on the chat (omit to preserve existing)
                personality_id:
                  type: string
                  format: uuid
                  description: Optional personality ID to set on the chat (omit to preserve existing; set to nil UUID to clear)
                tags:
                  type: array
                  maxItems: 10
                  items:
                    type: string
                    maxLength: 10
                  description: Optional chat tags (up to 10 tags, 10 chars each)
                is_favorite:
                  type: boolean
                  description: Optional favorite flag
                tts_autoplay_override:
                  type: boolean
                  nullable: true
                  description: Optional per-chat TTS autoplay override. Null means inherit personality default.
              required:
                - name
      responses:
        '200':
          description: Chat updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Patch chat
      description: Partially updates an existing chat (omitted fields are preserved)
      tags: [Chat Management]
      parameters:
        - name: id
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the chat session
                  maxLength: 200
                last_message_time:
                  type: string
                  format: date-time
                  description: Timestamp of the most recent message
                model_id:
                  type: string
                  format: uuid
                  description: Optional model ID to set on the chat
                personality_id:
                  type: string
                  format: uuid
                  description: Optional personality ID to set on the chat (set to nil UUID to clear)
                tags:
                  type: array
                  maxItems: 10
                  items:
                    type: string
                    maxLength: 10
                  description: Optional chat tags (up to 10 tags, 10 chars each)
                is_favorite:
                  type: boolean
                  description: Optional favorite flag
                archived:
                  type: boolean
                  description: When true, hides the thread from default lists; set to false to restore from the archive
                tts_autoplay_override:
                  type: boolean
                  nullable: true
                  description: Optional per-chat TTS autoplay override. Null means inherit personality default.
      responses:
        '200':
          description: Chat updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete chat
      description: Deletes a chat and all its messages
      tags: [Chat Management]
      parameters:
        - name: id
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Chat deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /chat/{id}/export:
    get:
      summary: Export chat
      description: |
        Exports a chat and all its messages as a ZIP file containing two files:
        - `chat.json`: Chat metadata (name, model, personality, timestamps)
        - `messages.jsonl`: All chat messages in JSON Lines format (one message per line)

        The export is streamed directly to the client for memory efficiency. The ZIP filename is based on the sanitized chat name.

        **Note**: If an error occurs during streaming, the ZIP file will be properly finalized but may be incomplete.
        The HTTP status code indicates whether the export was fully successful.
      tags: [Chat Management]
      parameters:
        - name: id
          in: path
          required: true
          description: Chat ID to export
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chat exported successfully
          content:
            application/zip:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: Attachment header with sanitized chat name as filename
              schema:
                type: string
                example: 'attachment; filename="My Chat.zip"'
            Cache-Control:
              description: Caching policy
              schema:
                type: string
                example: no-cache
        '400':
          description: Invalid chat ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Chat not found or user not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error during export. Note that if headers have already been sent, you may receive a valid but incomplete ZIP file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /webhooks/chat:
    get:
      summary: List threads (webhook)
      description: |
        Returns the token owner's threads, ordered by most recent activity. Requires the `chat:read` scope.

        Filter by persona with `personality_id`; `GET /webhooks/personality` lists the personas to choose from.
        Archived threads are excluded unless `archived=true`. Filters are validated: a malformed value is a 400, not ignored.
      tags: [Webhooks]
      security:
        - webhookAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page. Default 20; values above 100 are clamped to 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: personality_id
          in: query
          description: Only threads with this persona (see `GET /webhooks/personality`).
          schema:
            type: string
            format: uuid
        - name: name
          in: query
          description: Filter by thread name
          schema:
            type: string
        - name: search
          in: query
          description: Search in thread names and tags
          schema:
            type: string
        - name: tag
          in: query
          description: Filter by exact tag
          schema:
            type: string
        - name: is_favorite
          in: query
          description: Filter by favorite status
          schema:
            type: boolean
        - name: archived
          in: query
          description: When true, return only archived threads. When false or omitted, only active ones.
          schema:
            type: boolean
        - name: min_date
          in: query
          description: Only threads created at or after this time (RFC 3339).
          schema:
            type: string
            format: date-time
        - name: max_date
          in: query
          description: Only threads created at or before this time (RFC 3339).
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Threads retrieved successfully
          headers:
            Cache-Control:
              description: Always `no-store`; results are live state.
              schema:
                type: string
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/Chat'
        '400':
          description: Invalid paging or filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid webhook token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: The token does not have the `chat:read` scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login endpoint
    webhookAuth:
      type: http
      scheme: bearer
      description: Static webhook API token created via /webhook-tokens. Webhook calls act as the user who created the token, limited to the scopes the token was created with (`messages:write`, `chat:read`); a route that needs a scope the token lacks returns 403.
  schemas:
...
    JobStatus:
      type: string
      enum:
        - pending
        - processing
        - inference_complete
        - expression_complete
        - compaction_complete
        - complete
        - failed
      example: "pending"
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        job_type:
          type: string
        reference:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        error:
          type: string
        result_id:
          type: string
          format: uuid
        draft_deltas:
          type: array
          description: Incremental assistant text chunks emitted while inference is in progress.
          items:
            type: string
        draft_reasoning:
          type: array
          description: >-
            Incremental model reasoning chunks emitted while inference is in progress (z.ai GLM,
            Xiaomi MiMo). Display-only. Unlike draft_deltas this can be reset (emptied and
            re-sent) when a truncated call is retried, so clients should render the whole
            array on each poll rather than appending by cursor.
          items:
            type: string
        progress:
          type: string
          description: |
            Optional JSON-encoded progress payload for long-running jobs. Opaque to clients except per
            job_type. For `chat_import` it is `{phase, source, total, imported, skipped}`; for
            `account_import` it is `{phase, message, counts, conversations, personalities, memories,
            warnings, result}`, where `result` is present on terminal success. For `chat_message` it is
            the live tool-call timeline of the in-flight turn: `{tool_calls: [{id, name, input, status,
            output, round, started_at, finished_at}]}`, where `status` is `running`, `complete` or
            `error`, `input`/`output` are truncated previews, and `output`/`finished_at` are set once the
            call finishes. It is display-only; the saved assistant message's `tool_calls` are authoritative.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        user_id: "123e4567-e89b-12d3-a456-426614174001"
        job_type: "chat_message"
        reference: "msg_123"
        status: "processing"
        created_at: "2024-01-01T00:00:00Z"
        updated_at: "2024-01-01T00:00:00Z"

    # AgentJob Management Schemas
...
    Chat:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        unread_count:
          type: integer
          description: Number of unread assistant messages in this chat
        last_message_time:
          type: string
          format: date-time
        model_id:
          type: string
          format: uuid
        model_name:
          type: string
          description: Resolved display name of the associated model
        personality_id:
          type: string
          format: uuid
        personality_name:
          type: string
          description: Resolved display name of the associated personality
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 10
        is_favorite:
          type: boolean
        tts_autoplay_override:
          type: boolean
          nullable: true
          description: Optional per-chat TTS autoplay override. Null means inherit personality default.
        archived:
          type: boolean
          description: When true, the thread is in the archive and omitted from default list responses
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        user_id: "123e4567-e89b-12d3-a456-426614174001"
        name: "Research Project Discussion"
        last_message_time: "2024-01-01T12:00:00Z"
        tags: ["work", "todo"]
        is_favorite: false
        archived: false
        created_at: "2024-01-01T00:00:00Z"
        updated_at: "2024-01-01T00:00:00Z"
...
    PaginatedResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
        total_count:
          type: integer
        page:
          type: integer
        next_cursor:
          type: string
          description: >-
            Opaque keyset token for continuing a cursor-paginated view past the last row in this
            response (absent when there is nothing more). Only the descending chat-message view
            sets it.
      example:
        results: []
        total_count: 25
        page: 1
    Error:
      type: object
      properties:
        error:
          type: string
          deprecated: true
          description: |
            DEPRECATED — always identical to `message`. Retained only so existing clients
            keep working; read `message` and `code` instead. Removed once nothing reads
            it (ADR 0x020). Never contains internal error detail.
        message:
          type: string
          description: Human-readable explanation, safe to display to an end user
        code:
          $ref: '#/components/schemas/ErrorCode'
        details:
          description: Optional structured metadata for clients
          nullable: true
      required:
        - code
      example:
        error: "Personality name is required"
        message: "Personality name is required"
        code: "bad_request"
        details: null
    ErrorCode:
      type: string
      description: |
        Stable machine-readable error identifier. Always present on an error response.

        Codes come in two tiers (ADR 0x020):

        - **Specific** codes name a single failure condition and are a contract. Once
          published they are not renamed or repurposed, so clients may branch on them.
        - **Generic** codes are derived from the HTTP status and are the default for
          endpoints that have not been assigned a specific one yet. They are **not**
          stable branch targets: an endpoint may narrow from `bad_request` to a
          specific code at any time, and that is not a breaking change. Do not branch
          on a generic code — use the HTTP status instead.
      enum:
        # Generic — derived from HTTP status. Do not branch on these.
        - bad_request
        - unauthorized
        - forbidden
        - not_found
        - method_not_allowed
        - conflict
        - payload_too_large
        - unprocessable_entity
        - too_many_requests
        - internal_error
        - bad_gateway
        - service_unavailable
        - gateway_timeout
        - client_error
        - server_error
        # Specific — stable contracts, safe to branch on.
        - message_too_long
        - quota_exceeded
        - system_prompt_too_long

    # Cross-resource search schemas (GET /search)
...